+++ CHANGELOG.md
@@ -6,12 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- binsec branch to `operations/binsec`
 - Doctestable example for `model_predict` operation.
 - Doctestable examples to `operation/mapping.py`
+- shouldi got an operation to run Dependency-check on java code.
 - load function in high_level API
 - Doctestable examples to `db` operations.
+- Source for parsing `.ini` file formats
+- Tests for noasync high level API.
 ### Changed
 - `Edit on Github` button now hidden for plugins.
+- Doctests now run via unittests
 
 ## [0.3.7] - 2020-04-14
 ### Added